FEAT(#98): Improve Saved Recipes
