+++ src/css/App.css
@@ -112,19 +112,18 @@ nav.navbar .navbar-toggler-icon:after {
 }
 
 .badge {
-  position: absolute;
-  top: -10px;
-  right: 0;
   background-color: var(--color-primary);
   color: var(--color-secondary);
-  border-radius: 50%;
-  padding: 5px 10px;
+  border-radius: 12px;
+  padding: 2px 8px;
   font-size: 14px;
   font-weight: bold;
-  display: flex;
+  display: inline-flex;
   align-items: center;
   justify-content: center;
+  margin-left: 0.5rem;
 }
+
 
 /* Step 2 */
